Improve day 14 solution
Generalize the 4 different cases of tilting north/south/west/east with a
general function that tilts the platform in a given direction.

Use a linear single-pass algorithm to roll the rocks on a given line,
instead of the "Bubble Sort"-like algorithm used before. This is
basically changing a cubic O(n³) to a quadratic O(n²) with n=the side
length of the grid.

This change reduces the runtime of this solution from ~175ms to ~50ms :)